add des trucs qui marchent pas
