+++ src/View/AboutUs/index.html.twig
@@ -1,0 +1,23 @@
+{% extends 'layouts.html.twig' %}
+{% block title %} About Us {% endblock %}
+{% block stylesheet %}
+    <link rel="stylesheet" href="/assets/css/about_us.css">
+{% endblock %}
+{% block content %}
+<main>
+    <div id="mainPage">
+        <h1 id="pageTitle">L’équipe Blog-e-Boulga</h1>
+        <p id="pagePresentation">
+            Bienvenue dans notre équipe de développement web, la diversité de nos parcours professionnels se conjugue pour créer une synergie dans la réalisation de notre projet commun : un blog dédié au développement web.<br>
+            Notre objectif commun est de faire de cet espace une ressource pour les passionnés de développement web, des débutants aux professionnels chevronnés. Nous croyons que notre collaboration et nos perspectives variées contribueront à faire de ce projet une référence dans l'univers du développement web.
+        </p>
+        <figure class="separationAbout">
+            <img class="longSeparation" src="/assets/images/components/separation.svg" alt="separation">
+            <img class="smallSeparation" src="/assets/images/components/small_separation.svg" alt="small separation">
+        </figure>
+        {%for author in authors %}
+            {% include 'components/_about_us_card.html.twig' %}
+        {% endfor %}
+    </div>
+</main>
+{% endblock %}
