+++ src/View/components/_about_us_card.html.twig
@@ -1,0 +1,28 @@
+<article>
+    <div class="leftSide">
+        <figure class="profil">
+            <img src="assets/imges/authors/{{ author.imgSrc }}" alt="Lucas Boillot">
+        </figure>
+        <span class="authorName">
+                                <h2> {{ author.firstName }} . ' ' . {{ author.lastName }} </h2>
+                            </span>
+        <div class="website">
+            <a href={{ author.websiteURL }} ><img src="assets/images/components/myWebSite.svg" alt="Site web
+            personnel">
+                Mon Site</a>
+        </div>
+    </div>
+    <img class="firstquote" src="assets/images/components/quote.svg" alt="quote">
+    <aside>
+        <p>
+            {{ author.fullDescription }}  <span> <img src="assets/images/components/entypo_quote.svg" alt="endquote">
+            </span>
+        </p>
+    </aside>
+    <a class="linkedinProfil" href={{ author.linkedinUrl }} ><img src="assets/images/components/linkedin.svg"
+                                                                alt="Profil Linkedin"> Mon
+        profil Linkedin</a>
+    <a class="githubProfil" href={{ author.githubUrl }} ><img src="assets/images/components/github.svg" alt="Profil
+    Github"> Mon espace
+        Github</a>
+</article>
